Py3 division; convert to int

+++ p4/alignment_manip.py
@@ -2258,7 +2258,7 @@ if True:
         QB = 0.0
         zcount = 0                       # The number of double zeros in bigF
         # dof to start.  Half of off-diags of bigF
-        dof = ((dim * dim) - dim) / 2
+        dof = int(((dim * dim) - dim) / 2)
         for i in range(dim - 1):
             for j in range(i + 1, dim):
                 # Any double zeros?
@@ -2321,7 +2321,7 @@ if True:
             # The dof should be an int.
             PB = p4.func.chiSquaredProb(QB, dof2)
             PS = p4.func.chiSquaredProb(QS, dim - 1)
-            dof_R = (dim - 1) * (dim - 2) / 2
+            dof_R = int((dim - 1) * (dim - 2) / 2)
             PR = p4.func.chiSquaredProb(QR, dof_R)
             return (QB, QS, QR, PB, PS, PR)
         else:
